Initialize project with main functionality and VSCode extensions configuration
